revert tests due to them not working
